Update GoFigure - Clear operator boxes before solving the equation
Previously: If you enter any set of numbers & click "SOLVE" the answer
(operators in their respective boxes) were displayed. If you enter
another set of numbers which is not solveable, the previous answer
(operators in their respective boxes) is still visible.

Now: When "SOLVE" is clicked, the operator boxes are cleared before
solving so the previous answer is removed.